--- app/components/HeaderSimples.tsx
@@ -1,13 +1,0 @@
-export function Header() {
-    return (
-        <header>
-            <nav className="headerAQ navbar" style={{ backgroundColor: '#f9f9f9' }}>
-                <div className="container-fluid d-flex justify-content-start">
-                    <a className="navbar-brand d-flex align-items-center" href="/">
-                        <img src="/AprovaLogoHeader2Placeholder.png" alt="Logo" width="300" height="60" className="d-inline-block align-text-top me-2"></img>
-                    </a>
-                </div>
-            </nav>
-        </header>
-    );
-}
