+++ app/root.tsx
@@ -16,9 +16,8 @@ import main from "./styles/main.css?url";
 export const links: LinksFunction = () => [
   { rel: "stylesheet", href: main },
   { rel: "preconnect", href: "https://fonts.googleapis.com" },
-  { rel: "preconnect", href: "https://fonts.gstatic.com", crossOrigin: "anonymous" },
-  { rel: "stylesheet", href: "https://fonts.googleapis.com/css2?family=Inter:ital,opsz,wght@0,14..32,100..900;1,14..32,100..900&display=swap" },
-  { rel: "stylesheet", href: "https://cdnjs.cloudflare.com/ajax/libs/bootstrap-icons/1.5.0/font/bootstrap-icons.min.css" },
+  { rel: "preconnect", href: "https://fonts.gstatic.com", crossOrigin: "anonymous"},
+  { rel: "stylesheet", href: "https://fonts.googleapis.com/css2?family=Inter:ital,opsz,wght@0,14..32,100..900;1,14..32,100..900&display=swap"},
   { rel: "stylesheet", href: "https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css", integrity: "sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN", crossOrigin: "anonymous" },
   { rel: "icon", href: "/LogoAprovaQuiz.png", type: "image/png" },
 ];
@@ -36,7 +35,7 @@ export function Layout({ children }: { children: React.ReactNode }) {
         {children}
         <ScrollRestoration />
         <Scripts />
-        {/* <LiveReload /> */}
+        <LiveReload />
       </body>
     </html>
   );
@@ -49,9 +48,9 @@ export default function App() {
 
 export const meta: MetaFunction = ({ error }) => {
 
-  const typedError = error as { status: string }
+  const typedError = error as {status : string}
 
-  return [{ title: error ? `Error Status "${typedError.status}"` : "" }];
+  return [{ title: error? `Error Status "${typedError.status}"` : "" }];
 };
 
 export function ErrorBoundary() {
@@ -80,7 +79,6 @@ export function ErrorBoundary() {
         </>
       );
     } else if (error instanceof Error) {
-      console.log(error)
       return (
         <>
           <span>Error</span>
