--- app/styles/auth.css
@@ -1,132 +1,0 @@
-.auth-container {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    min-height: 80vh;
-    padding: 0 5%;
-    background-color: #f9f9f9;
-  }
-  
-  .auth-left {
-    flex: 1;
-    max-width: 50%;
-    padding-right: 2rem;
-  }
-  
-  .auth-right {
-    flex: 1;
-    max-width: 40%;
-    padding: 2rem;
-  }
-  
-  .auth-left h1 {
-    font-size: 3rem;
-    font-weight: 700;
-  }
-  
-  .auth-left p {
-    font-size: 1.2rem;
-  }
-  
-  .auth-right h2 {
-    font-size: 1.8rem;
-    margin-bottom: 1rem;
-    font-weight: bold;
-  }
- 
-  
-  .auth-right form {
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .auth-right input {
-    margin-bottom: 1rem;
-    padding: 1rem;
-    font-size: 1rem;
-    border: 2px solid #6f42dc; 
-    padding: 10px;
-    border-radius: 8px;
-    font-size: 16px;
-    width: 100%;
-    box-sizing: border-box;
-    transition: border-color 0.3s ease, background-color 0.3s ease;
-  }
-  
-  .auth-right input:focus {
-    background-color: #eae3f8; 
-    border-color: #5b32da; 
-    outline: none; 
-  }
-
-  .auth-right input::placeholder {
-    color: #5b32da; 
-  }
-
-  a {
-    color: #5b32da !important; 
-    text-decoration: none; 
-    transition: color 0.3s ease;
-  }
-  
-  a:hover {
-    color: #4b28b3; 
-    text-decoration: underline; 
-  }
-  
-  a:active {
-    color: #6f42dc; 
-  }
-  
-  .forgot-password {
-    color: #999999 !important;
-    text-decoration: none;
-    transition: color 0.3s ease;
-    margin-bottom: 1rem;
-  }
-  
-  .forgot-password:hover {
-    color: #666666 !important;
-    text-decoration: underline;
-  }
-  
-  .forgot-password:active {
-    color: #333333 !important; 
-  }
-  
-
-button {
-    background-color: #5b32da;
-    color: white; 
-    border: none;
-    padding: 10px 20px;
-    border-radius: 8px;
-    font-size: 16px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-  }
-  
-  button:hover {
-    background-color: #4b28b3; 
-  }
-  
-  button:active {
-    background-color: #6f42dc; 
-  }
-
-  input[type="date"] {
-    width: 100%;
-    padding: 10px;
-    margin: 10px 0;
-    border: 2px solid #6f42c1; 
-    background-color: white;
-    color: #6f42c1;
-  }
-
-  
-  input[type="date"]:focus {
-    background-color: #f3e8ff;
-    color: #6f42c1; 
-    outline: none;
-    border-color: #6f42c1; 
-  }
