--- app/styles/grid.css
@@ -1,11 +1,0 @@
-.container {
-    display: flex;
-    justify-content: center;
-    flex-wrap: wrap;
-    gap: 20px;
-    padding: 20px;
-  }
-
-.row {
-    gap: 20px;
-}
