+++ app/styles/home.css
@@ -1,242 +1,13 @@
 .bannerHome {
     background-color: var(--background-roxo-principal);
-    background: linear-gradient(0deg, rgba(162, 111, 235, 1) 0%, rgba(115, 54, 203, 1) 35%, rgba(138, 69, 237, 1) 100%);
+    background: linear-gradient(0deg, rgba(162,111,235,1) 0%, rgba(115,54,203,1) 35%, rgba(138,69,237,1) 100%); 
+    /*Depois tenho que trocar os rgb por variaveis*/
 }
 
-body {
-    font-family: 'League Spartan', sans-serif;
+.textoBanner{
+    color: var(--text-quaternario) !important;
 }
 
-h1, h2, h3, p, .botaoSimulado {
-    font-family: 'League Spartan', sans-serif;
-}
-
-h1 {
-    font-weight: 900;
-}
-
-.textaoBanner {
-    font-size: 56px;                  /* Tamanho da fonte (ajuste conforme necessário) */
-    line-height: 1.2;                 /* Alinhamento do texto */
-    margin-bottom: 1rem;              /* Distância abaixo do h1 */
-    color: #FFFFFF;    
-    font-weight: bolder;              /* Cor do texto (ajuste conforme necessário) */
-}
-
-.lead {
-    width: 690px;                     /* Largura fixa de 690 pixels */
-    height: 40px;                     /* Altura fixa de 40 pixels */
-    font-size: 28px;                  /* Tamanho da fonte */
-    line-height: 30px;                /* Para centralizar o texto verticalmente */
-    text-align: left;                 /* Alinhamento do texto à esquerda */
-    margin-top: 50px;                 /* Distância maior do topo (ajuste conforme necessário) */
-    font-weight: bolder;              /* Define o peso da fonte como bolder */
-    opacity: 0.5;                     /* Define a opacidade como 50% */
-}
-
-.textoBanner {
-    color: var(--text-quaternario) !important; /* Certifique-se de que a cor seja aplicada corretamente */
-    font-weight: bolder;               /* Define a fonte como negrito */
-}
-
-.botaoSimulado {
-    background-color: #FFFFFF;        /* Cor de fundo branca */
-    color: #3C1673;                   /* Cor do texto roxo */
-    border-radius: 80px;              /* Aumenta o border-radius */
-    border: 2px solid #3C1673;        /* Borda roxa */
-    width: 300px;                     /* Largura constante */
-    height: 80px;                     /* Aumenta a altura do botão */
-    font-size: 24px;                  /* Tamanho da fonte */
-    display: flex;                    /* Habilita o flexbox */
-    justify-content: center;          /* Alinha o texto horizontalmente ao centro */
-    align-items: center;              /* Alinha o texto verticalmente ao centro */
-    margin-top: 5px;                 /* Ajuste o valor conforme necessário */
-    text-decoration: none;            /* Remove sublinhado, se houver */
-}
-
-/* Estilo para a seta dentro do botão */
-.botaoSimulado .seta {
-    margin-left: 10px;                /* Espaçamento entre o texto e a seta */
-    color: #3C1673;                   /* Cor da seta */
-    font-size: 24px;                  /* Tamanho da seta (ajuste se necessário) */
-}
-
-.carousel {
-    position: relative;
-    width: 100%;
-    margin-bottom: 40px;
-}
-
-.carousel-item {
-    transition: transform 0.5s ease;
-}
-
-.card {
-    border-radius: 15px;
-    overflow: hidden;
-    box-shadow: 0px 5px 15px rgba(0, 0, 0, 0.2);
-    transition: transform 0.3s ease-in-out, box-shadow 0.3s ease-in-out; /* Adicionando transição para sombra */
-}
-
-.card:hover {
-    transform: scale(1.05);
-    box-shadow: 0px 10px 25px rgba(0, 0, 0, 0.3); /* Sombra mais pronunciada no hover */
-}
-
-.card img {
-    width: 100%;
-    height: 250px;
-    object-fit: cover;
-}
-
-.card-body {
-    padding: 15px;
-    height: 8rem;
-}
-
-.card-title {
-    font-size: 1.2rem;
-    font-weight: bold;
-    color: #3C1673;
-}
-
-.card-subtitle {
-    font-size: 1rem;
-    color: #4E596B;
-}
-
-.carousel-control-prev{
-    padding-right: 100px;
-}
-
-.custom-swiper-button {
-    position: absolute;
-    top: 50%;
-    transform: translateY(-50%);
-    z-index: 10;
-    color: #fff;
-    background-color: #3C1673;
-    border: none;
-    border-radius: 50%;
-    width: 40px;
-    height: 40px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    cursor: pointer;
-}
-
-.custom-swiper-button.prev {
-left: 10px;
-}
-
-.custom-swiper-button.next {
-right: 35px;
-}
-
-.custom-swiper-button:hover {
-background-color: rgba(162, 111, 235, 1);
-}
-
-
-.carousel-control-next, .carousel-control-prev {
-    width: 5% !important;
-}
-
-.carousel-indicators {
-    position: absolute;
-    bottom: -40px;
-    left: 50%;
-    transform: translateX(-50%);
-    display: flex;
-    justify-content: center;
-}
-
-.carousel-indicators li {
-    width: 12px;
-    height: 12px;
-    background-color: #3C1673;
-    border-radius: 50%;
-    margin: 0 5px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.carousel-indicators .active {
-    background-color: #FFD700; /* Cor dourada para a bolinha ativa */
-}
-
-/* PARTE DA NOTICIAS */
-.card-news {
-    border: none;
-    position: relative;
-    overflow: hidden;
-    transition: transform 0.3s ease-in-out;
-}
-
-.card-news:hover {
-    transform: scale(1.05);
-}
-
-.card-news .card-img-overlay {
-    position: absolute;
-    bottom: 0; /* Posiciona o texto no fundo do card */
-    left: 0; /* Posiciona o texto à esquerda */
-    width: 100%; /* Garante que o container use a largura completa */
-    padding: 10px; /* Dá um pequeno espaçamento para o texto */
-    background-color: rgba(0, 0, 0, 0.3); /* Fundo escuro transparente para melhor legibilidade */
-    box-sizing: border-box; /* Garante que o padding seja incluído nas dimensões do container */
-}
-
-.card-news .card-title {
-    font-size: 1.2rem; /* Tamanho menor do texto */
-    font-weight: bold;
-    color: white; /* Cor do texto */
-    margin: 0; /* Remove margem extra */
-    text-transform: uppercase;
-    text-align: left; /* Garante alinhamento à esquerda */
-}
-
-.card-news img {
-    width: 100%;
-    height: 300px;
-    object-fit: cover;
-}
-
-/* Consultas de Mídia para Responsividade */
-@media (max-width: 768px) {
-    .textaoBanner {
-        font-size: 36px;              /* Tamanho da fonte reduzido em telas menores */
-    }
-
-    .lead {
-        width: 100%;                  /* Largura ocupa 100% em telas menores */
-        font-size: 20px;              /* Tamanho da fonte reduzido */
-    }
-
-    .botaoSimulado {
-        width: 90%;                   /* Largura ocupa 90% em telas menores */
-        height: 60px;                 /* Altura do botão ajustada */
-        font-size: 20px;              /* Tamanho da fonte ajustado */
-    }
-}
-
-@media (max-width: 576px) {
-    .textaoBanner {
-        font-size: 28px;              /* Tamanho da fonte ainda mais reduzido */
-    }
-
-    .lead {
-        font-size: 18px;              /* Tamanho da fonte ajustado */
-    }
-
-    .botaoSimulado {
-        height: 50px;                 /* Altura do botão ajustada */
-        font-size: 18px;              /* Tamanho da fonte ajustado */
-    }
-}
-
-.icon{
-    width: 20px;
-    height: 20px;
+.botaoSimulado{
+    color: var(--text-primario) !important;
 }
