+++ app/styles/main.css
@@ -33,24 +33,6 @@
     text-decoration: none !important;
     color: var(--text-quaternario) !important;
 }
-
-.btn-outline-white {
-  color:  var(--text-quaternario) !important;
-  border: 1px solid  var(--text-quaternario) !important;
-  border-radius: 1.5rem !important; 
-}
-
-.btn-white {
-  background-color:  var(--text-quaternario) !important;
-  color: var(--background-roxo-principal) !important;
-  border: 1px solid var(--background-roxo-principal) !important;
-  border-radius: 2rem !important; 
-}
-
-.btn-white:hover {
-  background-color: #f0f0f0;
-}
-
 /* Fim do Header*/
 
 /* FOOTER*/
