--- app/styles/questoes.css
@@ -1,58 +1,0 @@
-.questao-bg {
-  background-color: var(--background-primario);
-}
-
-.questao-titulo {
-  color: var(--text-roxo-secundario) !important;
-  font-weight: bold;
-  font-size: 2rem;
-}
-
-.questao-texto {
-  font-size: 1.1rem;
-  margin-bottom: 1.5rem;
-}
-
-.opcoes {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.opcao {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-  padding: 0.5rem 1rem;
-  border-radius: 8px;
-  background-color: var(--background-primario);
-}
-
-.opcao-btn {
-  color: var(--text-roxo-principal);
-  border: var(--text-roxo-secundario) 5px solid;
-  padding: 0.5rem 1rem;
-  border-radius: 50%;
-  font-size: 1.2rem;
-  font-weight: bold;
-  cursor: pointer;
-}
-
-.opcao-marcada, .opcao-btn:hover {
-  color: #fff;
-  background-color: var(--text-roxo-principal);
-}
-
-.btn-responder {
-  background-color: var(--text-roxo-secundario);
-  color: var(--background-primario);
-  padding: 0.75rem 1.5rem;
-  border: none;
-  border-radius: 50px;
-  font-size: 1.1rem;
-  cursor: pointer;
-}
-
-.btn-responder:hover {
-  background-color: var(--text-roxo-principal);
-}
