--- app/styles/simpleCard.css
@@ -1,41 +1,0 @@
-.custom-card {
-    position: relative;
-    width: 100%;
-    height: 300px; 
-    overflow: hidden; 
-    border-radius: 15px; 
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-  }
-  
-  .card-background {
-    background-size: cover;
-    background-position: center;
-    width: 100%;
-    height: 100%;
-    position: relative;
-    border-radius: 15px;
-  }
-  
-  .gradient-overlay {
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    background: linear-gradient(to bottom, rgba(0, 0, 0, 0), rgba(0, 0, 0, 0.7));
-  }
-  
-  .card-title {
-    position: absolute;
-    bottom: 20px;
-    left: 20px;
-    color: white;
-    font-size: 1.5rem; /* Tamanho do título */
-    font-weight: bold;
-    text-shadow: 1px 1px 4px rgba(0, 0, 0, 0.7); /* Sombra no texto */
-  }
- 
-
-.buttonCard{
-  background: 0;
-  border: 0;
-}
