--- app/styles/simulado.css
@@ -1,45 +1,0 @@
-.h1Simulado {
-    color: var(--text-roxo-secundario) !important;
-    text-align: center;
-    padding-top: 3rem;
-    margin-bottom: 1rem;
-    font-weight: bold;
-}
-
-.txtSimulado {
-    color: var(--text-roxo-principal) !important;
-    display:flex;
-    justify-content: center;
-    margin-top: 2rem;
-    margin-bottom: 3rem;
-}
-
-.card-opcao {
-    background-color: var(--background-primario);
-    border-radius: 20px;
-  }
-  
-  .card-text{
-    color: var(--text-roxo-secundario) !important;
-    margin-bottom: 0.5rem;
-    font-size: 1.15rem;
-  }
-  
-  .card-number {
-    color: var(--text-roxo-secundario);
-    margin-bottom: 0.5rem;
-    font-weight: bold;
-  }
-
-  .col-6 {
-    max-width: 260px; 
-}
-
-.h-100 {
-    border-radius: 20px;
-}
-
-.btn{
- background-color: var(--text-roxo-principal);
- color: #fff;
-}
