+++ package.json
@@ -13,43 +13,33 @@
     "typegen": "wrangler types"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.6.0",
-    "@fortawesome/free-solid-svg-icons": "^6.6.0",
-    "@fortawesome/react-fontawesome": "^0.2.2",
-    "@remix-run/cloudflare": "^2.13.1",
+    "@remix-run/cloudflare": "^2.12.1",
     "@remix-run/cloudflare-pages": "^2.12.1",
-    "@remix-run/node": "^2.13.1",
-    "@remix-run/react": "^2.13.1",
-    "aprovaquiz": "file:",
-    "axios": "^1.7.7",
+    "@remix-run/react": "^2.12.1",
     "bootstrap": "^5.3.3",
-    "buffer": "^6.0.3",
-    "isbot": "^4.4.0",
-    "react": "^18.3.1",
+    "isbot": "^4.1.0",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.10.5",
-    "react-dom": "^18.3.1",
-    "sweetalert2": "^11.14.5",
-    "swiper": "^11.1.14",
-    "vite-plugin-node-polyfills": "^0.22.0"
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@cloudflare/workers-types": "^4.20241004.0",
+    "@cloudflare/workers-types": "^4.20240512.0",
     "@remix-run/dev": "^2.12.1",
-    "@types/react": "^18.3.11",
-    "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^6.21.0",
-    "@typescript-eslint/parser": "^6.21.0",
-    "autoprefixer": "^10.4.20",
-    "eslint": "^8.57.1",
-    "eslint-import-resolver-typescript": "^3.6.3",
-    "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-jsx-a11y": "^6.10.0",
-    "eslint-plugin-react": "^7.37.1",
-    "eslint-plugin-react-hooks": "^4.6.2",
-    "postcss": "^8.4.47",
-    "typescript": "^5.6.3",
-    "vite": "^5.4.8",
-    "vite-tsconfig-paths": "^4.3.2",
+    "@types/react": "^18.2.20",
+    "@types/react-dom": "^18.2.7",
+    "@typescript-eslint/eslint-plugin": "^6.7.4",
+    "@typescript-eslint/parser": "^6.7.4",
+    "autoprefixer": "^10.4.19",
+    "eslint": "^8.38.0",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "postcss": "^8.4.38",
+    "typescript": "^5.1.6",
+    "vite": "^5.1.0",
+    "vite-tsconfig-paths": "^4.2.1",
     "wrangler": "3.57.1"
   },
   "engines": {
