+++ vite.config.ts
@@ -1,7 +1,9 @@
-import { defineConfig } from 'vite';
-import tsconfigPaths from 'vite-tsconfig-paths';
-import { vitePlugin as remix, cloudflareDevProxyVitePlugin as remixCloudflareDevProxy } from '@remix-run/dev';
-import { nodePolyfills } from 'vite-plugin-node-polyfills'; // plugin para polyfills
+import {
+  vitePlugin as remix,
+  cloudflareDevProxyVitePlugin as remixCloudflareDevProxy,
+} from "@remix-run/dev";
+import { defineConfig } from "vite";
+import tsconfigPaths from "vite-tsconfig-paths";
 
 export default defineConfig({
   plugins: [
@@ -14,21 +16,5 @@ export default defineConfig({
       },
     }),
     tsconfigPaths(),
-    nodePolyfills({
-      // Configuração de polyfills
-      protocolImports: true, // Habilita polyfills automáticos para módulos de Node.js
-    }),
   ],
-  resolve: {
-    alias: {
-      // Fornecer polyfill para 'buffer'
-      buffer: 'buffer',
-    },
-  },
-  define: {
-    // Definir Buffer globalmente
-    global: {},
-    'process.env': {}, // Se necessário, pode definir outras variáveis de ambiente
-    Buffer: ['buffer', 'Buffer'],
-  },
 });
